requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
cairocffi==0.8.0          # via cairosvg, weasyprint
cairosvg==2.1.3           # via weasyprint
certifi==2018.1.18        # via requests
cffi==1.11.4              # via cairocffi, weasyprint
chardet==3.0.4            # via requests
cssselect2==0.2.1         # via cairosvg, weasyprint
defusedxml==0.5.0         # via cairosvg, python3-openid, social-auth-core
django-appconf==1.0.2     # via django-imagekit
django-imagekit==4.0.2
django-material==1.2.2
django-weasyprint==0.5.0
django==2.0.1
html5lib==1.0.1           # via weasyprint
idna==2.6                 # via requests
oauthlib==2.0.6           # via requests-oauthlib, social-auth-core
pdfrw==0.4                # via weasyprint
pilkit==2.0               # via django-imagekit
pillow==5.0.0
pycparser==2.18           # via cffi
pyjwt==1.5.3              # via social-auth-core
pyphen==0.9.4             # via weasyprint
python3-openid==3.1.0     # via social-auth-core
pytz==2017.3              # via django
pyyaml==3.12
requests-oauthlib==0.8.0  # via social-auth-core
requests==2.18.4          # via requests-oauthlib, social-auth-core
six==1.11.0               # via django-imagekit, html5lib, social-auth-app-django, social-auth-core
social-auth-app-django==2.1.0
social-auth-core==1.6.0   # via social-auth-app-django
tinycss2==0.6.1           # via cairosvg, cssselect2, weasyprint
urllib3==1.22             # via requests
weasyprint==0.42.2        # via django-weasyprint
webencodings==0.5.1       # via html5lib, tinycss2
